code/requirements.txt
pandas>=2.0.0
pydantic>=2.5.0
google-genai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
sentence-transformers
langchain-text-splitters
